requirements.txt
bcrypt~=4.0.1
chardet~=5.1.0
Pillow~=9.5.0
pygame~=2.4.0
requests~=2.30.0

psutil~=5.9.5